Update guide-to-the-documentation.md
I've suggested replacing the "Guide to the documentation" title with the "Learning Resources" one. Instead of providing users with just an overview of the docs, I think it makes sense to share more information about the available learning materials (and add a link to the Learn page), so that users that land directly in the tutorial can find them easily.